Split out model types from defs.
