+++ src/app/model-configuration/model-configuration.component.ts
@@ -1,6 +1,8 @@
+
 import { Component, OnInit } from '@angular/core';
 import { MatMenuTrigger } from '@angular/material/menu';
-import { modelSet, Model, riskProfiles } from '../model-defs';
+import { modelSet, riskProfiles } from '../model-defs';
+import { Model } from '../model';
 import { flattenHierarchy, FlatItem } from '../hierarchy';
 
 @Component({
